Fixed some warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@11625 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ graphics/ImageMap.h
@@ -324,8 +324,8 @@ const Colour4f ImageMap<V, VTraits>::vec3SampleTiled(Coord u, Coord v) const
 	int vt = i_pixels[1];
 	int ut_1 = wrapped_i_pixels_1[0];
 	int vt_1 = wrapped_i_pixels_1[1];
-	assert(ut >= 0 && ut < width && vt >= 0 && vt < height);
-	assert(ut_1 >= 0 && ut_1 < width && vt_1 >= 0 && vt_1 < height);
+	assert(ut >= 0 && ut < (int)width && vt >= 0 && vt < (int)height);
+	assert(ut_1 >= 0 && ut_1 < (int)width && vt_1 >= 0 && vt_1 < (int)height);
 
 	const Coord ufrac = frac[0];
 	const Coord vfrac = frac[1];
@@ -390,8 +390,8 @@ Map2D::Value ImageMap<V, VTraits>::sampleSingleChannelTiled(Coord u, Coord v, si
 	int vt = i_pixels[1];
 	int ut_1 = wrapped_i_pixels_1[0];
 	int vt_1 = wrapped_i_pixels_1[1];
-	assert(ut >= 0 && ut < width && vt >= 0 && vt < height);
-	assert(ut_1 >= 0 && ut_1 < width && vt_1 >= 0 && vt_1 < height);
+	assert(ut >= 0 && ut < (int)width && vt >= 0 && vt < (int)height);
+	assert(ut_1 >= 0 && ut_1 < (int)width && vt_1 >= 0 && vt_1 < (int)height);
 
 	const Coord ufrac = frac[0];
 	const Coord vfrac = frac[1];
